github_init/requirements.txt
click==8.0.1
Flask==2.1.2
gunicorn==20.1.0
itsdangerous==2.1.2
Jinja2==3.0.1
MarkupSafe==2.0.1
Werkzeug==2.1.2
line-bot-sdk==1.8.0
tensorflow-cpu==2.3.0